+++ docs/release_notes.md
@@ -12,7 +12,7 @@
 
 ### ✨ New Functionality
 
--
+- [Orchestration] [Added support to locally test prompt template files](https://sap.github.io/ai-sdk/docs/java/orchestration/chat-completion#locally-test-a-prompt-template)
 
 ### 📈 Improvements